#4 fix
Created data file and moved our dummy data into the data file.
Implemented EsLint in every file.
Removed Unused components.
