+++ reservacionVuelo/Screens/MyFlights.js
@@ -1,44 +1,11 @@
 import React, {Component} from 'react';
-import {View, Text, TouchableHighlight, ListView, FlatList} from 'react-native';
+import {View, Text, TouchableHighlight, FlatList} from 'react-native';
 import styles from '../components/styles';
 import Icon from 'react-native-vector-icons/FontAwesome';
 import {FlightCard} from '../components/flightCard';
+import dummyData from '../data/DummyData';
 
-const dummyData = [
-  {
-    id: 1,
-    departureCity: 'AGS',
-    departureCountry: 'MEXICO',
-    arrivalCity: 'LA',
-    arrivalCountry: 'USA',
-    year: 2021,
-    month: 'April',
-    day: 30,
-    passangers: 2,
-  },
-  {
-    id: 2,
-    departureCity: 'COL',
-    departureCountry: 'MEXICO',
-    arrivalCity: 'NY',
-    arrivalCountry: 'USA',
-    year: 2021,
-    month: 'April',
-    day: 22,
-    passangers: 3,
-  },
-  {
-    id: 3,
-    departureCity: 'MED ',
-    departureCountry: 'CDMX',
-    arrivalCity: 'BRL',
-    arrivalCountry: 'GERMANY',
-    year: 2021,
-    month: 'April',
-    day: 28,
-    passangers: 3,
-  },
-];
+
 
 export class MyFlights extends Component {
   constructor() {
